Payment 창 추가

+++ 2.Implement/DeliveryApp/app/src/main/java/com/project/deliveryapp/data/Stock.kt
@@ -12,6 +12,6 @@ data class Stock(
     val name: String,
     val releaseDate: LocalDateTime,
     val manufacturer: String,
-    val weight: Float,
-    val price: Long
+    val weight: Int,
+    val price: Int
 )
